requirements.txt
Django>=5.0
djangorestframework>=3.16.0
django-import-export>=4.3.0
pandas>=2.1.0
plotly>=6.3.0
python-dateutil>=2.8.0
reportlab>=4.0.0
openpyxl>=3.1.0
requests>=2.31.0
